refactor: 중복되는 tokenProvider 제거
