+++ hunmin-application/src/main/kotlin/com/application/auth/AuthService.kt
@@ -1,5 +1,6 @@
 package com.application.auth
 
+import com.common.global.auth.token.TokenProvider
 import com.common.global.exceptions.base.CustomException
 import com.common.util.throwWhen
 import com.domain.auth.Auth
@@ -9,14 +10,13 @@ import com.domain.auth.port.`in`.command.SignInCommand
 import com.domain.auth.port.`in`.command.SignUpCommand
 import com.domain.auth.port.out.AuthPasswordEncryptorPort
 import com.domain.auth.port.out.AuthRepositoryPort
-import com.domain.auth.port.out.TokenProviderPort
 import org.springframework.stereotype.Service
 
 @Service
 class AuthService(
     private val authRepositoryPort: AuthRepositoryPort,
     private val authPasswordEncryptorPort: AuthPasswordEncryptorPort,
-    private val tokenProviderPort: TokenProviderPort,
+    private val tokenProvider: TokenProvider,
 ) : AuthUseCase {
     override fun signUp(command: SignUpCommand): String {
         throwWhen(authRepositoryPort.existsByUsername(command.username)) {
@@ -32,7 +32,7 @@ class AuthService(
                 ),
             )
 
-        return tokenProviderPort.create(savedAuth.id)
+        return tokenProvider.create(savedAuth.id)
     }
 
     override fun signIn(command: SignInCommand): String {
@@ -44,6 +44,6 @@ class AuthService(
             AuthExceptionType.PASSWORD_INVALID_EXCEPTION
         }
 
-        return tokenProviderPort.create(auth.id)
+        return tokenProvider.create(auth.id)
     }
 }
